+++ examples/fs-imports/simple.proto
